requirement.txt
pathlib
pip install google-generativeai
python-dotenv
pip install youtube-transcript-api

streamlit
